--- wolfssl/wolfcrypt/wc_falcon_fpr.h
@@ -1,153 +1,0 @@
-/* wc_falcon_fpr.h
- *
- * Copyright (C) 2006-2026 wolfSSL Inc.
- *
- * This file is part of wolfSSL.
- *
- * wolfSSL is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 3 of the License, or
- * (at your option) any later version.
- *
- * wolfSSL is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
- */
-
-/*!
-    \file wolfssl/wolfcrypt/wc_falcon_fpr.h
-*/
-
-/* The Falcon floating-point primitive seam.
- *
- * Everything above this seam (FFT, Gaussian samplers, ffSampling, parts of
- * keygen) is written ONCE against the abstract fpr_* API declared here, and is
- * agnostic to the backend that implements it.
- *
- * Backends (selected at build time):
- *   - EMULATED (default): fpr is the IEEE-754 bit pattern carried in a word64;
- *     every operation is integer-only, fully deterministic and constant-time.
- *     This is the portable wolfCrypt default (no floating-point unit required).
- *     Implemented in wolfcrypt/src/wc_falcon_fpr.c.
- *   - NATIVE/ASM (opt-in, WOLFSSL_FALCON_FPR_ASM): fpr is a C double; operations
- *     map to per-architecture constant-time scalar FP (or DSP soft-float on
- *     ARMv7/Cortex-M). Generated by ../wolfssl-scripts.
- *
- * CONTRACT: every backend MUST produce results bit-identical to round-to-
- * nearest-even IEEE-754 binary64. The Gaussian sampler's determinism and
- * side-channel resistance depend on this. Hardware VDIV/VSQRT and any denormal
- * path are forbidden in the asm backends (data-dependent timing). */
-
-#ifndef WOLF_CRYPT_WC_FALCON_FPR_H
-#define WOLF_CRYPT_WC_FALCON_FPR_H
-
-#include <wolfssl/wolfcrypt/types.h>
-
-#if defined(HAVE_FALCON)
-
-#ifdef __cplusplus
-    extern "C" {
-#endif
-
-/* Backend selection. The emulated backend is the default; the native/asm
- * backend is opt-in and currently still carries the fpr value as a bit
- * pattern in a word64 so the seam type is uniform across translation units. */
-typedef word64 fpr;
-
-/* -- Constructors / conversions / arithmetic / predicates --------------- */
-
-/* Guard against selecting the native-double backend on a 32-bit ARM target that
- * has no double-precision FPU (Cortex-M0/M3/M4/M23/M33 are single-precision or
- * FPU-less). There, C double maps onto slow libgcc soft-double, so FPR_DOUBLE is
- * a pessimization -- the default emulated integer backend is faster. __ARM_FP
- * bit 3 (0x08) marks a hardware double FPU; AArch64 (which always has one) does
- * not define __arm__, so it is unaffected. */
-#if defined(WOLFSSL_FALCON_FPR_DOUBLE) && defined(__arm__) && \
-    (!defined(__ARM_FP) || (((__ARM_FP) & 0x08) == 0))
-    #warning "WOLFSSL_FALCON_FPR_DOUBLE on a target without a double-precision FPU falls back to slow soft-double; the default integer fpr backend is faster on Cortex-M."
-#endif
-
-#if defined(WOLFSSL_FALCON_FPR_DOUBLE)
-/* Inline native-double backend (opt-in). Maps the fpr seam onto the C double
- * type so the FFT/poly/sampler INLINE these scalar ops and keep values in FP
- * registers -- eliminating the per-op function call + GPR<->XMM shuffle that
- * dominate the out-of-line emulated/asm backends (~8x faster FP math, the bulk
- * of signing). Correctly-rounded IEEE-754 like the asm backend, with the same
- * constant-time-on-normals caveat: Falcon stays within normal range and the
- * caller must keep round-to-nearest-even (no FTZ/DAZ). fpr_expm_p63 and the fpr
- * constants still come from wc_falcon_fpr.c (which sees these inlines). */
-#include <math.h>
-static WC_INLINE double fpr__getd(fpr x) { double d; XMEMCPY(&d, &x, sizeof(d)); return d; }
-static WC_INLINE fpr fpr__setd(double d) { fpr x; XMEMCPY(&x, &d, sizeof(x)); return x; }
-static WC_INLINE fpr fpr_of(sword64 i)          { return fpr__setd((double)i); }
-static WC_INLINE fpr fpr_scaled(sword64 i, int sc) { return fpr__setd(ldexp((double)i, sc)); }
-static WC_INLINE sword64 fpr_rint(fpr x)        { double d = fpr__getd(x); return (sword64)llrint(d); }
-static WC_INLINE sword64 fpr_floor(fpr x)       { double d = floor(fpr__getd(x)); return (sword64)d; }
-static WC_INLINE sword64 fpr_trunc(fpr x)       { double d = trunc(fpr__getd(x)); return (sword64)d; }
-static WC_INLINE fpr fpr_add(fpr x, fpr y)      { return fpr__setd(fpr__getd(x) + fpr__getd(y)); }
-static WC_INLINE fpr fpr_sub(fpr x, fpr y)      { return fpr__setd(fpr__getd(x) - fpr__getd(y)); }
-static WC_INLINE fpr fpr_neg(fpr x)             { return fpr__setd(-fpr__getd(x)); }
-static WC_INLINE fpr fpr_half(fpr x)            { return fpr__setd(fpr__getd(x) * 0.5); }
-static WC_INLINE fpr fpr_double(fpr x)          { return fpr__setd(fpr__getd(x) + fpr__getd(x)); }
-static WC_INLINE fpr fpr_mul(fpr x, fpr y)      { return fpr__setd(fpr__getd(x) * fpr__getd(y)); }
-static WC_INLINE fpr fpr_sqr(fpr x)             { double d = fpr__getd(x); return fpr__setd(d * d); }
-static WC_INLINE fpr fpr_inv(fpr x)             { return fpr__setd(1.0 / fpr__getd(x)); }
-static WC_INLINE fpr fpr_div(fpr x, fpr y)      { return fpr__setd(fpr__getd(x) / fpr__getd(y)); }
-static WC_INLINE fpr fpr_sqrt(fpr x)            { return fpr__setd(sqrt(fpr__getd(x))); }
-static WC_INLINE int fpr_lt(fpr x, fpr y)       { return fpr__getd(x) < fpr__getd(y); }
-#else
-/* Convert a signed integer to fpr (exact for |i| < 2^53). */
-WOLFSSL_LOCAL fpr fpr_of(sword64 i);
-/* Convert i*2^sc to fpr. */
-WOLFSSL_LOCAL fpr fpr_scaled(sword64 i, int sc);
-/* Round to nearest integer (ties to even); toward -inf; toward zero. */
-WOLFSSL_LOCAL sword64 fpr_rint(fpr x);
-WOLFSSL_LOCAL sword64 fpr_floor(fpr x);
-WOLFSSL_LOCAL sword64 fpr_trunc(fpr x);
-WOLFSSL_LOCAL fpr fpr_add(fpr x, fpr y);
-WOLFSSL_LOCAL fpr fpr_sub(fpr x, fpr y);
-WOLFSSL_LOCAL fpr fpr_neg(fpr x);
-WOLFSSL_LOCAL fpr fpr_half(fpr x);
-WOLFSSL_LOCAL fpr fpr_double(fpr x);
-WOLFSSL_LOCAL fpr fpr_mul(fpr x, fpr y);
-WOLFSSL_LOCAL fpr fpr_sqr(fpr x);
-WOLFSSL_LOCAL fpr fpr_inv(fpr x);
-WOLFSSL_LOCAL fpr fpr_div(fpr x, fpr y);
-WOLFSSL_LOCAL fpr fpr_sqrt(fpr x);
-/* Returns 1 if x < y, else 0. Must be constant-time w.r.t. operand values. */
-WOLFSSL_LOCAL int fpr_lt(fpr x, fpr y);
-#endif /* WOLFSSL_FALCON_FPR_DOUBLE */
-
-/* -- Sampler support ---------------------------------------------------- */
-
-/* Compute, in fixed point (scaled by 2^63), ccs * exp(-x), for the
- * Gaussian-sampler Bernoulli test (BerExp). x and ccs are non-negative and
- * x stays in a bounded range guaranteed by the caller. */
-WOLFSSL_LOCAL word64 fpr_expm_p63(fpr x, fpr ccs);
-
-/* -- Named constants (defined by the active backend) -------------------- */
-
-extern const fpr fpr_zero;
-extern const fpr fpr_one;
-extern const fpr fpr_two;
-extern const fpr fpr_onehalf;
-extern const fpr fpr_invsqrt2;
-extern const fpr fpr_invsqrt8;
-extern const fpr fpr_ptwo31;       /*  2^31            */
-extern const fpr fpr_ptwo31m1;     /*  2^31 - 1        */
-extern const fpr fpr_mtwo31m1;     /* -(2^31 - 1)      */
-extern const fpr fpr_ptwo63m1;     /*  2^63 - 1        */
-extern const fpr fpr_mtwo63m1;     /* -(2^63 - 1)      */
-extern const fpr fpr_ptwo63;       /*  2^63            */
-
-#ifdef __cplusplus
-    }    /* extern "C" */
-#endif
-
-#endif /* HAVE_FALCON */
-#endif /* WOLF_CRYPT_WC_FALCON_FPR_H */
